refactor: 변경된 로직 서비스 전체에 적용
