+++ src/main/java/com/chungang/capstone/openstep/domain/Task/entity/Task.java
@@ -44,17 +44,6 @@ public class Task extends BaseEntity {
     }
 
     public void updatePrUrl(String prUrl) {
-        //url 형식이 올바른지 확인
-        //https://github.com/<owner>/<repo>/pull/<pr-number>
-        if (prUrl == null || !prUrl.matches("https://github\\.com/[^/]+/[^/]+/pull/\\d+")) {
-            //owner, repo 가 Task의 issue와 일치하는지 확인
-            String[] parts = prUrl.split("/");
-            if (parts.length < 5 || !parts[3].equals(issue.getRepo().getOwnerName()) || !parts[4].equals(issue.getRepo().getRepoName())) {
-                throw new TaskException(ErrorStatus.TASK_PR_URL_UPDATE_ERROR_INVALID_URL);
-            }
-
-        }
-
         if (this.status == TaskStatus.FORKED) {
             throw new TaskException(ErrorStatus.TASK_PR_URL_UPDATE_ERROR_EXCEPT_FORKED);
         }
